pyproject.toml
[build-system]
build-backend = "uv_build"
requires = [ "uv-build>=0.9.3,<0.10" ]

[project]
name = "mypackage"
version = "0.1.0"
description = "A research Python package."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.13"
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]

dependencies = [
  "hydra-core",
  "lightning",
  "numpy",
  "omegaconf",
  "rich",
  "timm",
  "torch",
  "torchmetrics",
  "torchvision",
  "wandb",
]

[dependency-groups]
dev = [
  "pre-commit>=4.3",
  "pytest>=8.4.2",
  "pytest-cov>=7",
  "ruff>=0.14.1",
  "ty>=0.0.1a23",
]

[tool.ruff]
target-version = "py313"
line-length = 100
force-exclude = true

lint.extend-select = [
  "ANN",  # flake8-annotations
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "EXE",  # flake8-executable
  "FA",   # flake8-future-annotations
  "FLY",  # flynt
  "FURB", # refurb
  "G",    # flake8-logging-format
  "I",    # isort
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "PERF", # Perflint
  "PGH",  # pygrep-hooks
  "PIE",  # flake8-pie
  "PT",   # flake8-pytest-style
  "PYI",  # flake8-pyi
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",
  "SIM",  # flake8-simplify
  "SLOT", # flake8-slots
  "TCH",  # flake8-type-checking
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle warnings
]
lint.ignore = [
  "ANN401",
  "COM812",
  "COM819",
  "D206",
  "D300",
  "E111",
  "E114",
  "E117",
  "PT011",
  "PT012",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "RET505",
  "RET506",
  "RUF005",
  "SIM108",
  "TRY003",
  # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
  "W191",
]
lint.extend-per-file-ignores."scripts/**" = [
  "ANN001",
  "ANN201",
]

lint.extend-per-file-ignores."tests/**" = [
  "ANN001",
  "ANN201",
  "RUF029",
  "SIM117",
  "SIM300",
]
lint.isort.known-first-party = [ "mypackage" ]

[tool.pytest.ini_options]
testpaths = [ "tests" ]
